BillingWrapperTests refactor (#617)
